requirements.txt
addict==2.4.0
asttokens==3.0.0
attrs==25.2.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.56.0
idna==3.10
importlib_metadata==8.6.1
ipython==8.34.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
narwhals==1.30.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.3
open3d==0.19.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
Pygments==2.19.1
pyparsing==3.2.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rerun-sdk==0.22.1
retrying==1.3.4
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.5.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
zipp==3.21.0
